camera-fraud-detection/requirements.txt
flask
opencv-python
exifread
pillow
numpy